+++ packages/babel-preset-expo/src/plugins/__tests__/babel-plugin-transform-export-namespace-from-upstream.test.ts
