+++ packages/babel-preset-expo/src/plugins/__tests__/expo-inline-manifest-plugin.test.ts
