+++ packages/babel-preset-expo/src/plugins/__tests__/expo-router-plugin.test.ts
