+++ packages/babel-preset-expo/src/plugins/__tests__/import-meta-transform-plugin.test.ts
@@ -1,6 +1,6 @@
 import * as babel from '@babel/core';
 
-import preset from '..';
+import preset from '../..';
 
 function getCaller(props: Record<string, string | boolean>): babel.TransformCaller {
   return props as unknown as babel.TransformCaller;
