+++ packages/babel-preset-expo/src/plugins/__tests__/use-dom-directive-plugin.test.ts
@@ -4,7 +4,7 @@
 
 import * as babel from '@babel/core';
 
-import preset, { type BabelPresetExpoOptions } from '..';
+import preset, { type BabelPresetExpoOptions } from '../..';
 
 const ENABLED_CALLER = {
   name: 'metro',
